Change LOTW filtering behavior in CR
Roster was treating LOTW filtering as "advisory" if someone had
previously set LoTW filtering to "show me nothing older than 3 months"
they would still get older stations but with "?" -- this is inconsistent
with every other filter, and is annoying--a station that once uploaded
to LoTW 8 years ago shouldn't be considered a LoTW station. Fixes this.